test(eth-jsonrpc): Added a python end-to-end testing subproject, integrated it into the normal testing flow, tested ethereum json-rpc.
This change introduces the infrastructure and testing framework for writing powerful high-level scenarios that test complex multi-node behaviors. It is based on ethereum json-rpc api (via web3.py client library), and at the moment the test scenarios there touch upon almost every method of the api.
The hope is that this will become a foundation for future work on end-to-end testing, as it is too tedious and slow to write this kind of tests in C++, plus this allows a better separation of labor, enabling dedicated testing engineers that need not know C++.